Fixes build warning for 32-bit Ubuntu
Implements: APPLINK-22313

+++ src/components/application_manager/src/commands/mobile/on_system_request_notification.cc
@@ -108,7 +108,7 @@ void OnSystemRequestNotification::AddHeader(BinaryMessage& message) const {
 
   char size_str[24];
 
-  if (0 > sprintf(size_str, "%lu", message.size())) {
+  if (0 > sprintf(size_str, "%zu", static_cast<size_t>(message.size()))) {
     memset(size_str, 0, sizeof(size_str));
   }
